+++ src/ArduinoJson/Collection/CollectionData.hpp
@@ -56,6 +56,7 @@ class CollectionData {
   void remove(VariantSlot *slot);
 
   size_t memoryUsage() const;
+  size_t nesting() const;
   size_t size() const;
 
  private: