+++ src/ArduinoJson/Object/ObjectRef.hpp
@@ -46,6 +46,10 @@ class ObjectRefBase {
     return _data ? _data->memoryUsage() : 0;
   }
 
+  FORCE_INLINE size_t nesting() const {
+    return _data ? _data->nesting() : 0;
+  }
+
   FORCE_INLINE size_t size() const {
     return _data ? _data->size() : 0;
   }